fixed double exception error

+++ PIM/src/main/java/presentation/EditProductCommand.java
@@ -26,7 +26,7 @@ public class EditProductCommand extends Command {
        String productnamedescription = (request.getParameter("ProductNameDescription"));
        String productdescription = (request.getParameter("ProductDescription"));
        String companyname = (request.getParameter("CompanyName"));
-       double price = Integer.parseInt(request.getParameter("Price"));
+       double price = Double.parseDouble(request.getParameter("Price"));
        int quantity = Integer.parseInt(request.getParameter("Quantity"));
        String picturename = (request.getParameter("PictureName"));
        String minorcategory = "1";
